requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
openai==1.109.1
python-dotenv==1.1.1
pandas==2.3.2
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.3.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
requests==2.32.3
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.2.0